made the mouse emulation delay parameter for touch interactions configurable via the Settings page

+++ src/gui/Utilities/config.cpp
@@ -28,9 +28,9 @@ static const auto* CONFIG_TOUCH_DOWN_DELAY = "touch_down_delay";
 static const auto* CONFIG_TOUCH_UP_DELAY = "touch_up_delay";
 static const auto *CONFIG_TOUCH_DRAG_START_DELAY = "touch_drag_start_delay";
 
-static constexpr uint64_t default_touch_down_delay = 50;
-static constexpr uint64_t default_touch_up_delay = 50;
-static constexpr uint64_t default_touch_drag_start_delay = 150;
+static constexpr uint32_t default_touch_down_delay = 50;
+static constexpr uint32_t default_touch_up_delay = 50;
+static constexpr uint32_t default_touch_drag_start_delay = 150;
 
 using namespace fsmapper;
 using namespace nlohmann;
@@ -53,9 +53,9 @@ class config_imp : public config{
     uint64_t lua_standard_libraries{MOPT_STDLIB_BASE | MOPT_STDLIB_PACKAGE | MOPT_STDLIB_MATH | MOPT_STDLIB_TABLE | MOPT_STDLIB_STRING};
     std::string skipped_version{"0.0"};
     dcs_exporter_mode dcs_exporter_mode_data{dcs_exporter_mode::unknown};
-    uint64_t touch_down_delay{default_touch_down_delay};
-    uint64_t touch_up_delay{default_touch_up_delay};
-    uint64_t touch_drag_start_delay{default_touch_drag_start_delay};
+    uint32_t touch_down_delay{default_touch_down_delay};
+    uint32_t touch_up_delay{default_touch_up_delay};
+    uint32_t touch_drag_start_delay{default_touch_drag_start_delay};
 
     template <typename KEY, typename VALUE>
     void reflect_number(json& jobj, const KEY& key, VALUE& var){
@@ -255,22 +255,22 @@ public:
     void set_dcs_exporter_mode(dcs_exporter_mode value){
         update_value(dcs_exporter_mode_data, value);
     }
-    uint64_t get_touch_down_delay(){
+    uint32_t get_touch_down_delay(){
         return touch_down_delay;
     }
-    void set_touch_down_delay(uint64_t value){
+    void set_touch_down_delay(uint32_t value){
         update_value(touch_down_delay, value);
     }
-    uint64_t get_touch_up_delay(){
+    uint32_t get_touch_up_delay(){
         return touch_up_delay;
     }
-    void set_touch_up_delay(uint64_t value){
+    void set_touch_up_delay(uint32_t value){
         update_value(touch_up_delay, value);
     }
-    uint64_t get_touch_drag_start_delay(){
+    uint32_t get_touch_drag_start_delay(){
         return touch_drag_start_delay;
     }
-    void set_touch_drag_start_delay(uint64_t value){
+    void set_touch_drag_start_delay(uint32_t value){
         update_value(touch_drag_start_delay, value);
     }
     void reset_touch_delay(){
